+++ core/Forge/src/main/java/de/markusbordihn/easynpc/client/ClientPlayerEventHandler.java
@@ -21,7 +21,7 @@ package de.markusbordihn.easynpc.client;
 
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.eventbus.api.listener.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 
 @EventBusSubscriber(value = Dist.CLIENT)